requirements.txt
python-freeDictionaryAPI
httpx
